+++ CHANGES.rst
@@ -5,8 +5,10 @@ Changelog
 4.2b17 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- Added `MeetingItem.validate_pollType` that relies on
+  `ChangeItemPollTypeView.validate_new_poll_type` to make sure that it is not
+  possible to break encoded votes from the item edit form.
+  [gbastien]
 
 4.2b16 (2021-09-28)
 -------------------